Use preferred common name instead of English
I initially chose English-centric because I was seeing inconsistent
results from the API. But now I'm not sure if that was because I was
testing with an unauthenticated request -- a quick test now shows that
changing my language on the site changes an authenticated request's
response language.

Let's give this a go.