[sync] Download in parallel to observation processing
Download takes a significant amount of the wall time of sync, but is a
very obvious candidate for running in parallel, since it is network
heavy.

Unfortunately the Lightroom coroutines are a bit crippled, but it's not
too hard to start a background task to download while continuing to
yield observations in the iterator, and wait for it to complete if we
run out of observations to yield.

In my measurements this reduces the synchronization time by around 20%.
My tests were not thorough, and I only ran comparison measurements once
or twice, but overall I found what I saw to be convincing enough that
this technique was worthwhile.