Switch to dkjson library for JSON decode/encode
 - It's faster
 - It has a smaller code size
 - It has the same license as everything else in this repo

The main motivation was speed during sync. The list of observations is
large, and JSON parsing took a fair amount of time (1.5s per 100
observations), which can not be parallelized as it is pure CPU. dkjson
is about twice as fast at parsing these large objects as the previously
used library, in my tests.

This change reduced synchronization time by around 20% on my machine.