Upload photo before creating observation
The photo upload API returns a `to_observation` field, which is an
observation object derived from the photo. Using this is better than
doing our own derivation because it ensures consistency between this
plugin and other means of uploading to iNaturalist. For example, the way
timezones are handled if the camera is incorrectly configured currently
differs.

bf61c02a35736c720f91d5802d22e232686f4c4c did the opposite of this
change; the thinking was that we could always create an observation
photo, so POST to create the observation without any photos specified
would not be destructive. In experimenting and examining the iNaturalist
Rails code, I found that a POST to /observations with photos specified
would actually add to the photos if the observation already existed, so
this path is safe.

This also removes the option to upload keywords, since that is
controlled by the Lightroom metadata options.