Check Image Sizing settings on save
iNaturalist limits image sizes to 2048px on the long edge, and uploading
larger images will just take longer, for no good reason.

While we can't restrict how the user configures the Image Sizing dialog,
we can easily examine it after the fact, and present suggestions.

I tried to LrView.bind() to the settings to see if that would be a good
model, but was unable to observe changes to the relevant keys.