Return nil from API :new if no accessToken
This will cause errors in other parts of the program to be thrown
earlier, for "attempt to index local 'api' (a nil value)". Importantly,
it happens early enough that the progress meter is not set up yet, and
so the user is not stuck in a state where they can't close the sync.

The error message is still not pretty, but this is a path that should
not happen -- an access token should always be available.

Fixes #14