Check access token during UI login verification
When opening a catalog on a different computer, the access token stored
with LrPasswords will not be available. Without this change, there is no
way to get an access token from this state -- the "Log in" button is
greyed out, as there is a username stored in the plugin propertyTable of
the catalog.

This change allows the user to log in again when moving their catalog to
a new computer.